POIs

Metadata dictionary

POIs.create_poi_metadata — Function

create_poi_metadata(osm_filename::String, json_filename::String, dir::String = "datasets")::Dict{String, Dict{String, String}}

Auxilary function - it returns a dictionary of metadata which is used as an argument in several other functions. Arguments:

osm_filename - an .osm filename located in dir directory

json_filename - a JSON file where the types and subtypes of POIs are configured (check datasets/POI_config.json as an example)

dir - a directory where the the JSON file is located and where the .osm files are located.

Description of the returned metadata dictionary:

The keys of the dictionary are the names of temporary files which are used to obtain POIs using function osm_to_dict from src/osm_parser.

The values are dictionaries of metadata for each of the temporary files, in the following form:

`Dict("primary_type" => primary_type, "subtype" => subtype, "city" => city, "input_filepath" => input_filepath, "osm_query" => osm_query, "output_filepath" => output_filepath)`

The dictionary keys have the following meanings:

primary_type - primary_type taken from the JSON file

subtype - subtype taken from the JSON file

city - city name extracted from the name of the .osm file located in the 'dir' directory (e.g. if the file name is 'Beijing.osm', then the city is 'beijing')

input_filepath - it is used in osm_to_dict to locate the .osm file, which is processed with osm_filter

osm_query - this is a query to obtain the temporary file (e.g. --keep=amenity=school) from which POIs are extracted using function osm_to_dict from src/osm_parser.

output_filepath - this is the path to the temporary file which is created to extract the POIs from .osm file

Parsing .osm file

POIs.delete_version_tags! — Method

delete_version_tags!(dict::Dict{AbstractString, AbstractString})::Dict{String, String}

Auxilary function used in osm_to_dict to parse .osm file.

POIs.dict_of_attributes — Function

dict_of_attributes(c::LightXML.XMLElement, name::String = LightXML.name(c))::Dict{String, String}

Auxilary function used in osm_to_dict to parse .osm file.

POIs.process_attributes — Method

process_attributes(dict::Dict{String, String})::Dict{String, Union{Int, String}}

Auxilary function used in osm_to_dict to parse .osm file.

POIs.assign_attr_to_poi_object! — Method

assign_attr_to_poi_object!(poi::POIObject, attr::Dict{String, String})

Auxilary function used in osm_to_dict to parse .osm file.

POIs.generate_temporary_file — Method

generate_temporary_file(filename::String, metadata::Dict{String, Dict{String, String}})

Auxilary function - it generates a temporary file for further processing and returns a filepath of this file Arguments:

filename - name of the temporary file thich is to be generated

metadata - a metadata generated from function create_poi_metadata from src/poi_metadata.jl

POIs.osm_to_dict — Function

osm_to_dict(filename::String, metadata::Dict{String, Dict{String, String}}, excluded_keywords::Array{String} = ["text", "bounds"])::Dict{String, Vector{POIObject}}

Auxilary function - parses .osm file and returns a dictionary whose key is a name of a parsed temporary file, and value is a vector of parsed POIs. A single POI is represented as a POIObject type which is a mutable struct with fields defined in src/types.jl. Arguments:

filename - the name of the .osm file that the function parses (e.g. beijing.osm)

metadata - a metadata generated from function createpoimetadata from src/poi_metadata.jl

excluded_keywords - keywords excluded from parsing

Vector of POIs

POIs.get_coordinates_of_way — Method

get_coordinates_of_way(object_data::Vector{POIObject}, way::POIObject)::Dict{String, Float64}

Auxilary function used inside get_coordinates function.

POIs.get_coordinates — Method

get_coordinates(object_data::Vector{POIObject}, element::POIObject)::Dict{String, Float64}

Auxilary function - it returns lat and lon coordinates of a POI object. If not found, then they are 0. Arguments:

object_data - a vector of POI objects in which an element is located

element - a POI object for which the coordinates are to be found

POIs.get_data_vector — Method

get_data_vector(metadata::Dict{String, Dict{String, String}})::Vector{Dict{String, Vector{POIObject}}}

Auxilary function - it returns a vector of dictionaries - each of them being generated using osm_to_dict function from src/osm_parser.jl. The number of elements of dictionary depends on the metadata. Arguments:

metadata - metadata dictionary generated using function create_poi_metadata from src/poi_metadata.jl

POIs.get_poi_types — Method

get_poi_types(metadata::Dict{String, Dict{String, String}})::Tuple{Vector{String}, Vector{String}}

Auxilary funtion - it returns a tuple of vectors - each vector representing primary_types or subtypes extrcted from a metadata dictionary. Arguments:

metadata - metadata dictionary generated using function createpoimetadata from src/poi_metadata.jl

POIs.create_poi_dataset — Method

create_poi_dataset(object_data::Dict{String, Vector{POIObject}}, primary_type::String, subtype::String)::Vector{ProcessedPOI}

Auxilary function - it returns a processed dataset (vector of elements of type ProcessedPOI) with the POIs of one type (primary_type and subtype). Arguments:

object_data - it is a raw parsed set of POI objects (output of osm_to_dict)

primary_type - this is a primary type that will be assigned to the processed POIs

subtype - this is a subtype that will be assigned to the processed POIs

POIs.generate_poi_vectors — Function

generate_poi_vectors(osm_filename::String, poi_config::String = "POI_config.json")

High level function - returns the vector of processed poi datasets. Each dataset is from a different type and subtype and is represented by a vector of processed POIs. Arguments:

osm_filename - name of .osm file from which the POIs are processed and generated

poi_config - a JSON file with configuration of the POIs that are to be generated.

The function works in the following way step by step:

It creates the metadata for a desired .osm file, based on JSON dictionary with config.

It creates a vector of raw datasets for each of the files from metadata. The datasets are generated using the osm_to_dict function from src/osm_parser.jl.

It creates vectors of primary_types and subtypes based on the metadata.

It transforms each raw dataset (each element of the vector) to the processed dataset with POis using a function generate_poi_dataset.

Creating dataframe of POIs (one POI type)

POIs.columns — Method

columns(processed_objects::Vector{ProcessedPOI})::Vector{String}

Auxilary function - it returns a vector of all distinct osm tag keys which are used as colnames of the df. Arguments:

processed_objects - vector of processed POIs of one type (output of create_poi_dataset function)

POIs.create_df — Function

create_df(processed_objects::Vector{ProcessedPOI}, df_columns::Vector{String} = String[])::DataFrame

Auxilary function - it returns the dataframe with processed POIs of one primary_type and one subtype Arguments:

processed_objects - vector of processed pois of one type (output of create_poi_dataset)

df_columns - vector of column names for the dataframe (output of columns function)

Creating a dataframe of all POIs

POIs.columns_in_poi_vector — Method

columns_in_poi_vector(processed_objects_vector::Vector{Vector{ProcessedPOI}})::Vector{String}

Auxilary function - it returns a vector of all distinct osm tag keys which are used as column names of the df. Arguments:

processed_objects_vector - vector of processed pois of all types (output of generatepoivectors)

POIs.create_poi_df — Method

create_poi_df(processed_objects_vector::Vector{Vector{ProcessedPOI}})::DataFrame

Main function - it returns the dataframe of all POIs of all configured primary_types and subtypes Arguments:

processed_objects_vector - the vector of processed pois of all types (output of generate_poi_vectors)

Filtering dataframe columns

POIs.filter_columns_by_threshold — Function

filter_columns_by_threshold(dframe::DataFrame, threshold::Float64 = 0.5)::DataFrame

Main function - it filters columns of the poi dataframe and returns a dataframe with those columns, whose fraction of non-missing values exceeds the threshold value Arguments:

dframe - a DataFrame with POIs

threshold - a minimum fraction of non-missing values in a column

POIs.filter_columns_by_colnames — Function

filter_columns_by_colnames(dframe::DataFrame, colnames::Vector{String} = String[])::DataFrame

Main function - it filters columns of the poi dataframe and returns dataframe with defined colnames Arguments:

dframe - a DataFrame with POIs

colnames - vector of columns in output dataframe