OSMgetPOI

Downloading files

OSMgetPOI.download_bbbike_file — Method

download_bbbike_file(url::String; target_filepath::String = "datasets/file")

Main function - it downloads a .gz file from bbbike and unzips it. Returns a name for the unzipped file. Arguments:

url - url link to the .gz file on bbbike website

target_filepath - the target path for the downloaded and unzipped file. If default, the file will be called file.osm and save in datasets directory.

OSMgetPOI.download_geofabrik_file — Method

download_geofabrik_file(url::String; target_filepath::String = "datasets/file")

Main function - it downloads a .bz2 file from geofabrik and unzips it. Returns a name for the unzipped file. Arguments:

url - url link to the .bz2 file on geofabrik website

target_filepath - the target path for the downloaded and unzipped file. If default, the file will be called file.osm and save in datasets directory.

Parsing .osm file

OSMgetPOI.delete_version_tags! — Method

delete_version_tags!(dict::Dict{AbstractString, AbstractString})::Dict{String, String}

Auxilary function used in osm_to_dict to parse .osm file.

OSMgetPOI.dict_of_attributes — Function

dict_of_attributes(c::LightXML.XMLElement, name::String = LightXML.name(c))::Dict{String, String}

Auxilary function used in osm_to_dict to parse .osm file.

OSMgetPOI.process_attributes — Method

process_attributes(dict::Dict{String, String})::Dict{String, Union{Int, String}}

Auxilary function used in osm_to_dict to parse .osm file.

OSMgetPOI.assign_attr_to_poi_object! — Method

assign_attr_to_poi_object!(poi::POIObject, attr::Dict{String, String})

Auxilary function used in osm_to_dict to parse .osm file.

OSMgetPOI.generate_temporary_file — Method

generate_temporary_file(osm_filename::String, poi_type::POITypes.POIType)

Auxilary function - it generates a temporary file for further processing and returns a filepath of this file Arguments:

osm_filename - name of the temporary file thich is to be generated

poi_type - POIType for which the temporary file is to be genarated

OSMgetPOI.osm_to_dict — Function

osm_to_dict(osm_filename::String, poi_type::POITypes.POIType, excluded_keywords::Array{String} = ["text", "bounds"])::Dict{POITypes.POIType, Dict{Int, POIObject}}

Auxilary function - parses .osm file and returns a dictionary whose key is a ::POIType and value is a vector of parsed POIs. A single POI is represented as a POIObject type which is a mutable struct with fields defined in src/types.jl. Arguments:

osm_filename - the name of the .osm file that the function parses (e.g. beijing.osm)

poi_type - a POIType for which the file is parsed

excluded_keywords - keywords in .osm file excluded from parsing. Suggested to use the default.

Vector of POIs

OSMgetPOI.get_coordinates_of_way — Method

get_coordinates_of_way(object_data::Dict{Int, POIObject}, way_id::Int)::Dict{String, Float64}

Auxilary function used inside get_coordinates function. Arguments:

object_data - a vector of POI objects in which an element is located

way - a way for which one wants to find a lat-lon coordinates, represented by POIObject.

OSMgetPOI.get_coordinates — Method

get_coordinates(object_data::Dict{Int, POIObject}, poi_id::Int)::Dict{String, Float64}

Auxilary function - it returns lat and lon coordinates of a POI object. If not found, then they are 0. Arguments:

object_data - a vector of POI objects in which an element is located

element - a POI object for which the coordinates are to be found

OSMgetPOI.delete_duplicated_elements! — Method

delete_duplicated_elements!(processed_poi_dict::Dict{Int, ProcessedPOI}, poi::POIObject, data::Dict{Int, POIObject})

Auxilary function - it mutates processedpoidict. Previously, the poi of type POIObject was transformed into ProcessedPOI and added to the processedpoidict. If the processedpoidict already has elements that are child nodes of poi, then these child elements are deleted from processedpoidict. Arguments:

processed_poi_dict - a vector of ProcessedPOIs

poi - a POIObject which we are transorming into ProcessedPOI.

data - a POIobject dataset

OSMgetPOI.create_poi_dataset — Method

create_poi_dataset(object_data::Dict{POITypes.POIType, Dict{Int, POIObject}})::Vector{ProcessedPOI}

Auxilary function - it returns a processed dataset (vector of elements of type ProcessedPOI) with the POIs of one POItype. Arguments:

object_data - it is a raw parsed set of POIObjects (output of osm_to_dict)

OSMgetPOI.generate_poi_vectors — Method

generate_poi_vectors(osm_filename::String, poi_types::POITypes.POIType...)::Vector{Vector{ProcessedPOI}}

High level function - returns the vector of processed poi datasets. Each dataset is of a different POIType defined in the function arguments and is represented by a vector of ProcessedPOIs. Arguments:

osm_filename - name of .osm file from which the POIs are processed and generated

poi_types - all POITypes, for which the dataframe should be generated

Creating dataframe of POIs (one POI type)

OSMgetPOI.columns — Method

columns(processed_objects::Vector{ProcessedPOI})::Vector{String}

Auxilary function - it returns a vector of all distinct osm tag keys which are used as colnames of the dataframe. Arguments:

processed_objects - vector of processed POIs of one poitype (output of `createpoi_dataset` function)

OSMgetPOI.create_df — Function

create_df(processed_objects::Vector{ProcessedPOI}, df_columns::Vector{String} = String[])::DataFrame

Auxilary function - it returns the dataframe with processed POIs of one POIType Arguments:

processed_objects - vector of processed pois of one poi_type

df_columns - vector of column names for the dataframe (output of columns function)

Creating a dataframe of all POIs

OSMgetPOI.columns_in_poi_vector — Method

columns_in_poi_vector(processed_objects_vector::Vector{Vector{ProcessedPOI}})::Vector{String}

Auxilary function - it returns a vector of all distinct osm tag keys which are used as column names of the df. Arguments:

processed_objects_vector - vector of processed pois of all poitypes (output of generatepoi_vectors)

OSMgetPOI.create_poi_df — Function

create_poi_df(processed_objects_vector::Vector{Vector{ProcessedPOI}}, threshold::Float64 = 1.0, columns::Vector{String} = ["addr:housenumber", "addr:street", "addr:postcode", "addr:country"])::DataFrame

Auxilary function - it returns the dataframe of all POIs of all configured POITypes Arguments:

processed_objects_vector - the vector of processed pois of all POITypes

threshold - a minimum fraction of non-missing values in a column

columns - vector of columns that are to be included in the returned dataframe, regardless of the threshold value

Creating a dataframe of all POIs from .osm file

OSMgetPOI.get_poi_df — Method

get_poi_df(osm_filename::String, poi_types::POITypes.POIType ...; columns::Vector{String} = ["addr:housenumber", "addr:street", "addr:postcode", "addr:country"], threshold::Float64 = 1.0)

Main function - it returns the dataframe of all POIs of all configured POITypes from the .osm file.

Arguments:

osm_filename - path to .osm file from which the POIs are processed and generated

threshold - a minimum fraction of non-missing values in a column

poi_types - all POITypes, for which the dataframe should be generated

columns - vector of columns that are to be included in the returned dataframe, regardless of the threshold value

Filtering dataframe columns

OSMgetPOI.filter_columns — Function

filter_columns(dframe::DataFrame, threshold::Float64 = 1.0, columns::Vector{String} = ["addr:housenumber", "addr:street", "addr:postcode", "addr:country"])

Main function - it filters columns of the poi dataframe and returns a dataframe with those columns, whose fraction of non-missing values exceeds the threshold value Arguments:

dframe - a DataFrame with POIs

threshold - a minimum fraction of non-missing values in a column

columns - vector of columns that are to be included in the returned dataframe, regardless of the threshold value