Creating a metadata dictionary for other functions

POIs.create_poi_metadata — Function

create_poi_metadata(osm_filename::String, json_filename::String, dir::String = "datasets")

Auxilary function - it returns a dictionary of metadata which is used as an argument in several other functions. Arguments:

jsonfilename - a JSON file where the types and subtypes of POIs are configured (check datasets/POIconfig.json as an example)

dir - a directory where the the JSON file is located and where the .osm files are located.

Description of returned metadata dictionary: The keys of the dictionary are the names of temporary files which are used to obtain POIs using function osmtodict from src/osmparser. The values are dictionaries of metadata for each of the temporary files, in the following form: Dict("primarytype" => primarytype, "subtype" => subtype, "city" => city, "inputfilepath" => inputfilepath, "osmquery" => osmquery, "outputfilepath" => output_filepath) The dictionary keys have the following meanings:

primarytype - primarytype taken from the JSON file

subrype - subtype taken from the JSON file

city - city name extracted from the name of the .osm file located in the 'dir' directory (e.g. if the file name is 'Beijing.osm', then the city is 'beijing')

inputfilepath - it is used in osmtodict to locate the .osm file, which is processed with osmfilter

osmquery - this is a query to obtain the temporary file (e.g. –keep=amenity=school) from which POIs are extracted using function osmtodict from src/osmparser.

output_filepath - this is the path to the temporary file which is created to extract the POIs from .osm file

Parsing .osm file

POIs.generate_temporary_file — Method

generate_temporary_file(filename::String, metadata::Dict{String, Dict{String, String}})

Auxilary function - it generates a temporary file for further processing and returns a filepath of this file Arguments:

filename - name of the temporary file thich is to be generated

metadata - a metadata generated from function createpoimetadata from src/poi_metadata.jl

POIs.osm_to_dict — Function

osm_to_dict(filename::String, metadata::Dict{String, Dict{String, String}}, excluded_keywords::Array{String} = ["text", "bounds"])::Dict{String, Vector{Dict{String, Any}}}

Auxilary function - parses .osm file and returns a dictionary whose key is a name of a parsed temporary file, and value is a vector of parsed POIs. A single POI is represented as a POIObject type which is a mutable struct with fields defined in src/types.jl. Arguments:

filename - the name of the .osm file that the function parses (e.g. beijing.osm)

metadata - a metadata generated from function createpoimetadata from src/poi_metadata.jl

excluded_keywords - keywords excluded from parsing

POIs.delete_version_tags! — Method

delete_version_tags!(dict::Dict{AbstractString, AbstractString})::Dict{String, String}

Auxilary function used in osmtodict to parse .osm file.

POIs.dict_of_attributes — Function

dict_of_attributes(c::LightXML.XMLElement, name::String = LightXML.name(c))::Dict{String, String}

Auxilary function used in osmtodict to parse .osm file.

POIs.process_attributes — Method

process_attributes(dict::Dict{String, String})::Dict{String, Union{Int, String}}

Auxilary function used in osmtodict to parse .osm file.

POIs.assign_attr_to_poi_object! — Method

assign_attr_to_poi_object!(poi::POIObject, attr::Dict{String, String})

Auxilary function used in osmtodict to parse .osm file.

Creating a vector of POIs for a selected city

POIs.get_coordinates_of_way — Method

get_coordinates_of_way(object_data::Vector{POIObject}, way::POIObject)::Dict{String, Float64}

Auxilary function used inside get_coordinates function.

POIs.get_coordinates — Method

get_coordinates(object_data::Vector{POIObject}, element::POIObject)::Dict{String, Float64}

Auxilary function - it returns lat and lon coordinates of a POI object. If not found, then they are 0. Arguments:

object_data - a vector of POI objects in which an element is located

element - a POI object for which the coordinates are to be found

POIs.get_data_vector — Method

get_data_vector(metadata::Dict{String, Dict{String, String}})::Vector{Dict{String, Vector{POIObject}}}

Auxilary function - it returns a vector of dictionaries - each of them being generated using osmtodict function from src/osm_parser.jl. The number of elements of dictionary depends on the metadata. Arguments:

metadata - metadata dictionary generated using function createpoimetadata from src/poi_metadata.jl

POIs.get_poi_types — Method

get_poi_types(metadata::Dict{String, Dict{String, String}})::Tuple{Vector{String}, Vector{String}}

Auxilary funtion - it returns a tuple of vectors - each vector representing primary_types or subtypes extrcted from a metadata dictionary. Arguments:

metadata - metadata dictionary generated using function createpoimetadata from src/poi_metadata.jl

POIs.create_poi_dataset — Method

create_poi_dataset(object_data::Dict{String, Vector{POIObject}}, primary_type::String, subtype::String)::Vector{ProcessedPOI}

Auxilary function - it returns a processed dataset (vector of elements of type ProcessedPOI) with the POIs of one type (primary_type and subtype). Arguments:

objectdata - it is a raw parsed set of POI objects (output of osmto_dict)

primary_type - this is a primary type that will be assigned to the processed POIs

subtype - this is a subtype that will be assigned to the processed POIs

POIs.generate_poi_vectors — Function

generate_poi_vectors(osm_filename::String, poi_config::String = "POI_config.json")

High level function - returns the vector of processed poi datasets. Each dataset is from a different type and subtype and is represented by a vector of processed POIs. Arguments:

osm_filename - name of .osm file from which the POIs are processed and generated

poi_config - a JSON file with configuration of the POIs that are to be generated.

The function works in the following way step by step:

It creates the metadata for a desired .osm file, based on JSON dictionary with config.

It creates a vector of raw datasets for each of the files from metadata.

The datasets are generated using the osmtodict function from src/osm_parser.jl.

It creates vectors of primary_types and subtypes based on the metadata.

It transforms each raw dataset (each element of the vector) to the processed dataset with POis using a function generatepoidataset.

Creating dataframe from one vector of processed POIs (only one POI type)

POIs.columns — Method

columns(processed_objects::Vector{ProcessedPOI})::Vector{String}

Auxilary function - it returns a vector of all distinct osm tag keys whichare used as colnames of the df. Arguments:

processedobjects - vector of processed POIs of one type (output of createpoi_dataset function)

POIs.create_df — Function

create_df(processed_objects::Vector{ProcessedPOI}, df_columns::Vector{String} = String[])::DataFrame

Auxilary function - it returns the dataframe with processed POIs of one primary_type and one subtype Arguments:

processedobjects - vector of processed pois of one type (output of createpoi_dataset)

df_columns - vector of column names for the dataframe (output of columns function)

Creating a dataframe from all vectors of POIs (output of generatepoivectors function)

POIs.columns_in_poi_vector — Method

columns_in_poi_vector(processed_objects_vector::Vector{Vector{ProcessedPOI}})::Vector{String}

Auxilary function - it returns a vector of all distinct osm tag keys which are used as column names of the df. Arguments:

processedobjectsvector - vector of processed pois of all types (output of generatepoivectors)

POIs.create_poi_df — Method

create_poi_df(processed_objects_vector::Vector{Vector{ProcessedPOI}})::DataFrame

Main function - it returns the dataframe of all POIs of all primary_types and subtypes Arguments:

processedobjectsvector - the vector of processed pois of all types (output of generatepoivectors)

Filtering dataframe columns

POIs.filter_columns — Function

filter_columns(dframe::DataFrame, threshold::Float64 = 0.5)::DataFrame

Main function - it filters columns of the poi dataframe and returns a dataframe with those columns, whose fraction of non-missing values exceeds the threshold value Arguments:

dframe - a DataFrame with POIs

threshold - a minimal fraction of non-missing values in a column